run: add check for etcd service
A true value means that this node is an active participant in the etcd cluster.
This will bind the etcd service to etcd-announce so that the dns service
discovery can work.

A false value means that this node is just a proxy participant of the etcd
cluster. No one cares to reach this node externally so no need to announce.
Setup etcd to be in proxy mode.