lochness: setup hosts file
The hosts file includes the nodes ip address and fqdn as per queensland.
This is needed because there are some racy conditions where etcd will not
be able to lookup a node (itself) that the dns discovery protocol says it
should be able to find. It can not find it because queensland is no longer
talking to etcd.