etcd: do not enable PROXY mode
This conflicts with first-boot of etcd-nodes since the initial (single node)
cluster on each node will not have `/config/etcd` set `true` and kappa will
dutifully make this node a proxy. Later kappa will try to fix that once
the data is correctly restored but now etcd craps out and we cannot finish the
cluster. Fix is found and will be tested/committed soon.