tftpd: replace in.tfptd with dnsmasq
tftpd is dying without any info, when it does run it fails to serve up files
and if it did serve files then there was a good chance of transmission errors.
On the other hand dnsmasq implements only the minimum needed for net-boot and
has given no errors. Life is too short, using dnsmasq.